+++ NEWS.md
@@ -3,6 +3,7 @@ PENDING: dataiter 0.27
 
 * `DataFrame.__getattr__`: Fix error message when attribute not found
 * `DataFrame.read_csv`: Try to avoid mixed types
+* `Vector.rank`: Rewrite, add `method` argument
 * `*.read_*`: Rename fname argument path
 * `*.write_*`: Rename fname argument path